# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediERP - Medical Store ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: parentId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account (only if no entries)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}/balance:
    get:
      operationId: getAccountBalance
      tags: [accounts]
      summary: Get account balance and ledger
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Account balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountBalance"

  # ── JOURNAL ENTRIES ───────────────────────────────────────
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI chat about the medical store data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAssistantResponse"
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, type, code, isSystem, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        code: { type: string }
        parentId: { type: ["integer", "null"] }
        parentName: { type: ["string", "null"] }
        isSystem: { type: boolean }
        isActive: { type: boolean }
        balance: { type: number }
        description: { type: ["string", "null"] }
        createdAt: { type: string }
        children:
          type: array
          items: { $ref: "#/components/schemas/Account" }
    AccountInput:
      type: object
      required: [name, type, code]
      properties:
        name: { type: string }
        type:
          type: string
          enum: [asset, liability, equity, revenue, expense]
        code: { type: string }
        parentId: { type: integer }
        description: { type: string }
    AccountUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        description: { type: string }
        isActive: { type: boolean }
    AccountBalance:
      type: object
      required: [accountId, accountName, balance, debits, credits]
      properties:
        accountId: { type: integer }
        accountName: { type: string }
        balance: { type: number }
        debits: { type: number }
        credits: { type: number }
        transactions:
          type: array
          items: { $ref: "#/components/schemas/LedgerLine" }
    LedgerLine:
      type: object
      required: [id, date, description, debit, credit, balance]
      properties:
        id: { type: integer }
        date: { type: string }
        description: { type: string }
        reference: { type: ["string", "null"] }
        debit: { type: number }
        credit: { type: number }
        balance: { type: number }

    # ── JOURNAL ───────────────────────────────────────────────
...
    AiChatInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
        history:
          type: array
          items: { $ref: "#/components/schemas/AiMessage" }
    AiMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content: { type: string }
    AiAssistantResponse:
      type: object
      required: [response]
      properties:
        response: { type: string }